Finalizing the project in its initial form

+++ src/main/java/com/example/task_system/user/repository/UserRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepo extends JpaRepository<Users,Integer> {
+public interface UserRepo extends JpaRepository<Users,Long> {
 
     boolean existsByEmailIgnoreCase(String email);
